SS-9456: Separate parameter types from realization
