+++ entities/parameter/config/theme/parameterSelectionComponentTheme.ts
@@ -1,0 +1,19 @@
+import type {MantineThemeComponent} from "@mantine/core";
+import type {InteractionEffect} from "@shapediver/viewer.shared.types";
+
+export interface ParameterSelectionComponentStyleProps {
+	selectionColor?: InteractionEffect;
+	availableColor?: InteractionEffect;
+	hoverColor?: InteractionEffect;
+}
+
+export type ParameterSelectionComponentThemePropsInput =
+	Partial<ParameterSelectionComponentStyleProps>;
+
+export function ParameterSelectionComponentThemeProps(
+	props: ParameterSelectionComponentThemePropsInput,
+): MantineThemeComponent {
+	return {
+		defaultProps: props,
+	};
+}
